requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.36.0
SQLAlchemy==2.0.31
psycopg[binary]==3.2.9
alembic==1.16.5
pydantic-settings==2.4.0
Jinja2==3.1.6
python-dotenv>=1.0
aiogram == 3.21.0
openai == 1.108.1
langid == 1.1.6
PyYAML==6.0.2
Telethon==1.41.2
starlette==0.47.2
qrcode==8.2
django-qrcode==0.3
pillow==11.3.0
bcrypt==4.3.0
passlib==1.7.4
itsdangerous>=2.2,<3
python-multipart==0.0.19
authlib==1.6.4
auto-i18n-lib==0.1.3
requests==2.32.5
pathspec==0.12.1
sse-starlette==3.0.2